+++ src/main/java/com/project/Karman/dto/request/MatchQuarterCreateRequestDto.java
@@ -16,7 +16,7 @@ public record MatchQuarterCreateRequestDto(
 
         @NotNull(message = "실점은 필수 입력 값입니다.")
         @PositiveOrZero(message = "실점은 0 이상이어야 합니다.")
-        Integer concededGoal,
+        Long concededGoal,
 
         @NotNull(message = "라인업은 필수 입력 값입니다.")
         @NotEmpty(message = "라인업은 최소 1명 이상이어야 합니다.")
